Swapped the threading model
