+++ Extensions/WikiPDFExport/WikiPDFExportTask/test/testrunnerForWikiWithNoClone.ts
@@ -1,9 +1,5 @@
 import {
-    CloneWikiRepo
-    } from "../src/GitWikiFuntions";
-
-import {
-    ExportPDF
+    ExportRun
     } from "../src/ExportFunctions";
 
 function logInfo (msg: string) {
@@ -14,8 +10,28 @@ function logError (msg: string) {
     console.log("\x1b[31m", msg);
 }
 
-const wikiPath = `${__dirname}\\..\\..\\..\\..\\..\\..\\AzurePipelines.wiki`;
+const localpath = `${__dirname}\\..\\..\\..\\..\\..\\..\\AzurePipelines.wiki`;
 const outputFile = "c:\\tmp\\test\\output.pdf";
+const injectExtraHeader = false;
+const singleFile = "";
+const extraParams = "";
+const useAgentToken = false;
+const cloneRepo = false;
+const user = "";
+const password = "";
+const repo = "";
+const branch = "";
 
-console.log(`Current folder is ${__dirname}`);
-ExportPDF (wikiPath, "" , outputFile, "", logInfo, logError);
+ExportRun(
+    cloneRepo,
+    localpath,
+    singleFile,
+    outputFile,
+    extraParams,
+    useAgentToken,
+    repo,
+    user,
+    password,
+    injectExtraHeader,
+    branch
+);
